Refactor user profile page to support static export
